updating before starting the addition of video version

+++ code/analysis/run_analyze_human_model.py
@@ -39,7 +39,7 @@ if __name__ == '__main__':
     parser.add_argument('-m', '--model_names', type=str, nargs='+')
     parser.add_argument('-word_model', '--word_model_name', type=str, default='fasttext')
     parser.add_argument('-window_size', '--window_size', type=int, default=25)
-    parser.add_argument('-prosody', '--prosody', type=int, default=0)
+    parser.add_argument('-careful_whisper', '--careful_whisper', type=int, default=0)
     parser.add_argument('-o', '--overwrite', type=int, default=0)
     p = parser.parse_args()
 
@@ -96,38 +96,48 @@ if __name__ == '__main__':
 
     # Analyze results from cleaned data 
     df_results = pd.read_csv(os.path.join(results_dir, f'task-{p.task}_group-cleaned-behavior.csv'))
-    df_analyzed_results = analysis.analyze_human_results(df_transcript, df_results, word_model_info, window_size=p.window_size, top_n=None, drop_rt=None)
-
-    # Add wrong phoneme info
-    df_analyzed_results = add_wrong_phoneme_info(df_results, df_analyzed_results)
-
-    # Add in prosody columns and add to the list
     out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_human.csv')
-    df_analyzed_results.loc[:, prosody_columns] = df_transcript.loc[df_analyzed_results['word_index'], prosody_columns].reset_index(drop=True)
-    df_analyzed_results.to_csv(out_fn, index=False)
-    
-    ### Now do model data
-    df_all_models = []
 
-    for model_name in p.model_names:
-        # Load model data and add wrong phoneme information
-        df_model = analysis.analyze_model_accuracy(df_transcript, word_model_info=word_model_info, models_dir=models_dir, model_name=model_name, task=p.task, window_size=p.window_size, candidate_rows=candidate_rows, lemmatize=False)
-        df_model = add_wrong_phoneme_info(df_results, df_model)
+    if not os.path.exists(out_fn) or p.overwrite:
+        df_analyzed_results = analysis.analyze_human_results(df_transcript, df_results, word_model_info, window_size=p.window_size, top_n=None, drop_rt=None)
+
+        # Add wrong phoneme info
+        df_analyzed_results = add_wrong_phoneme_info(df_results, df_analyzed_results)
 
         # Add in prosody columns and add to the list
-        df_model.loc[:, prosody_columns] = df_transcript.loc[df_model['word_index'], prosody_columns].reset_index(drop=True)
-        df_all_models.append(df_model)
 
-    # Concatenate models and then add to the humans dataframe
-    df_all_models = pd.concat(df_all_models).reset_index(drop=True)
+        df_analyzed_results.loc[:, prosody_columns] = df_transcript.loc[df_analyzed_results['word_index'], prosody_columns].reset_index(drop=True)
+        df_analyzed_results.to_csv(out_fn, index=False)
+    else:
+        df_analyzed_results = pd.read_csv(out_fn)
+    
+    ### Now do model data
+    if p.careful_whisper:
+        out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_window-size-{str(p.window_size).zfill(5)}_human-careful-whisper.csv')
+    else:
+        out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_window-size-{str(p.window_size).zfill(5)}_human-model.csv')
 
-    df_combined = pd.concat([
-        df_analyzed_results.loc[:, combined_columns], 
-        df_all_models.loc[:, combined_columns]
-    ]).reset_index(drop=True)
+    if not os.path.exists(out_fn) or p.overwrite:
+        df_all_models = []
 
-    out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_window-size-{str(p.window_size).zfill(5)}_human-model.csv')
-    df_combined.to_csv(out_fn, index=False)
+        for model_name in p.model_names:
+            # Load model data and add wrong phoneme information
+            df_model = analysis.analyze_model_accuracy(df_transcript, word_model_info=word_model_info, models_dir=models_dir, model_name=model_name, task=p.task, window_size=p.window_size, candidate_rows=candidate_rows, lemmatize=False)
+            df_model = add_wrong_phoneme_info(df_results, df_model)
+
+            # Add in prosody columns and add to the list
+            df_model.loc[:, prosody_columns] = df_transcript.loc[df_model['word_index'], prosody_columns].reset_index(drop=True)
+            df_all_models.append(df_model)
+
+        # Concatenate models and then add to the humans dataframe
+        df_all_models = pd.concat(df_all_models).reset_index(drop=True)
+
+        df_combined = pd.concat([
+            df_analyzed_results.loc[:, combined_columns], 
+            df_all_models.loc[:, combined_columns]
+        ]).reset_index(drop=True)
+
+        df_combined.to_csv(out_fn, index=False)
 
     ########################################################
     ########## Repeat process for lemmatized words #########
@@ -135,38 +145,46 @@ if __name__ == '__main__':
 
     # Analyze human lemmatized data
     df_lemmatized_results = pd.read_csv(os.path.join(results_dir, f'task-{p.task}_group-cleaned-behavior_lemmatized.csv'))
-    df_analyzed_lemmatized = analysis.analyze_human_results(df_transcript, df_lemmatized_results, word_model_info, window_size=p.window_size, top_n=None, drop_rt=None)
-
-    # Add wrong phoneme info
-    df_analyzed_results = add_wrong_phoneme_info(df_lemmatized_results, df_analyzed_lemmatized)
-
-    # Add in prosody columns and add to the list
     out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_human-lemmatized.csv')
-    df_analyzed_lemmatized.loc[:, prosody_columns] = df_transcript.loc[df_analyzed_lemmatized['word_index'], prosody_columns].reset_index(drop=True)
-    df_analyzed_lemmatized.to_csv(out_fn, index=False)
+    
+    if not os.path.exists(out_fn) or p.overwrite:
+        df_analyzed_lemmatized = analysis.analyze_human_results(df_transcript, df_lemmatized_results, word_model_info, window_size=p.window_size, top_n=None, drop_rt=None)
 
-    ### Now do model data
-    df_lemmatized_models = []
-
-    for model_name in p.model_names:
-        df_model = analysis.analyze_model_accuracy(df_transcript, word_model_info=word_model_info, models_dir=models_dir, model_name=model_name, task=p.task, window_size=p.window_size, candidate_rows=candidate_rows, lemmatize=True)
-        df_model = add_wrong_phoneme_info(df_lemmatized_results, df_model)
+        # Add wrong phoneme info
+        df_analyzed_results = add_wrong_phoneme_info(df_lemmatized_results, df_analyzed_lemmatized)
 
         # Add in prosody columns and add to the list
-        df_model.loc[:, prosody_columns] = df_transcript.loc[df_model['word_index'], prosody_columns].reset_index(drop=True)
-        df_lemmatized_models.append(df_model)
+        df_analyzed_lemmatized.loc[:, prosody_columns] = df_transcript.loc[df_analyzed_lemmatized['word_index'], prosody_columns].reset_index(drop=True)
+        df_analyzed_lemmatized.to_csv(out_fn, index=False)
+    else:
+        df_analyzed_lemmatized = pd.read_csv(out_fn)
 
-    # Concatenate models and then add to the humans dataframe
-    df_lemmatized_models = pd.concat(df_lemmatized_models).reset_index(drop=True)
+    ### Now do model data
+    if p.careful_whisper:
+        out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_window-size-{str(p.window_size).zfill(5)}_human-careful-whisper-lemmatized.csv')
+    else:
+        out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_window-size-{str(p.window_size).zfill(5)}_human-model-lemmatized.csv')
 
-    out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_window-size-{str(p.window_size).zfill(5)}_human-model-lemmatized.csv')
+    if not os.path.exists(out_fn) or p.overwrite:
+        df_lemmatized_models = []
 
-    df_lemmatized_combined = pd.concat([
-        df_analyzed_lemmatized.loc[:, combined_columns], 
-        df_lemmatized_models.loc[:, combined_columns]
-    ]).reset_index(drop=True)
+        for model_name in p.model_names:
+            df_model = analysis.analyze_model_accuracy(df_transcript, word_model_info=word_model_info, models_dir=models_dir, model_name=model_name, task=p.task, window_size=p.window_size, candidate_rows=candidate_rows, lemmatize=True)
+            df_model = add_wrong_phoneme_info(df_lemmatized_results, df_model)
 
-    df_lemmatized_combined.to_csv(out_fn, index=False)
+            # Add in prosody columns and add to the list
+            df_model.loc[:, prosody_columns] = df_transcript.loc[df_model['word_index'], prosody_columns].reset_index(drop=True)
+            df_lemmatized_models.append(df_model)
+
+        # Concatenate models and then add to the humans dataframe
+        df_lemmatized_models = pd.concat(df_lemmatized_models).reset_index(drop=True)
+
+        df_lemmatized_combined = pd.concat([
+            df_analyzed_lemmatized.loc[:, combined_columns], 
+            df_lemmatized_models.loc[:, combined_columns]
+        ]).reset_index(drop=True)
+
+        df_lemmatized_combined.to_csv(out_fn, index=False)
 
     ########################################################
     ########### Conduct distribution comparison ############
@@ -174,20 +192,21 @@ if __name__ == '__main__':
 
     df_lemmatized_results = pd.read_csv(os.path.join(results_dir, f'task-{p.task}_group-cleaned-behavior_lemmatized.csv'))
 
-    df_all_comparisons = []
-
-    for model_name in p.model_names:
-        print (model_name)
-
-        df_comparison = analysis.compare_human_model_distributions(df_lemmatized_results, word_model_info, models_dir=logits_dir, model_name=model_name, task=p.task, lemmatize=True)
-        df_comparison = add_wrong_phoneme_info(df_lemmatized_results, df_comparison)
-        df_comparison.loc[:, prosody_columns] = df_transcript.loc[df_comparison['word_index'], prosody_columns].reset_index(drop=True)
-        df_all_comparisons.append(df_comparison)
-
-    if p.prosody:
-        out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_window-size-{str(p.window_size).zfill(5)}_human-prosody-model-distributions-lemmatized.csv')
+    if p.careful_whisper:
+        out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_window-size-{str(p.window_size).zfill(5)}_human-careful-whisper-distributions-lemmatized.csv')
     else:
         out_fn = os.path.join(results_dir, f'task-{p.task}_group-analyzed-behavior_window-size-{str(p.window_size).zfill(5)}_human-model-distributions-lemmatized.csv')
     
-    df_all_comparisons = pd.concat(df_all_comparisons).reset_index(drop=True)
-    df_all_comparisons.to_csv(out_fn, index=False)
+    if not os.path.exists(out_fn) or p.overwrite:
+        df_all_comparisons = []
+
+        for model_name in p.model_names:
+            print (model_name)
+
+            df_comparison = analysis.compare_human_model_distributions(df_lemmatized_results, word_model_info, models_dir=logits_dir, model_name=model_name, task=p.task, lemmatize=True)
+            df_comparison = add_wrong_phoneme_info(df_lemmatized_results, df_comparison)
+            df_comparison.loc[:, prosody_columns] = df_transcript.loc[df_comparison['word_index'], prosody_columns].reset_index(drop=True)
+            df_all_comparisons.append(df_comparison)
+
+        df_all_comparisons = pd.concat(df_all_comparisons).reset_index(drop=True)
+        df_all_comparisons.to_csv(out_fn, index=False)
